More serialization fixups
