+++ src/xunit.execution.win8/xunit.execution.win8.csproj
@@ -81,6 +81,9 @@
     <Compile Include="..\common\TestOptionsNames.cs">
       <Link>Common\TestOptionsNames.cs</Link>
     </Compile>
+    <Compile Include="..\common\XunitSerializationInfo.cs">
+      <Link>Common\XunitSerializationInfo.cs</Link>
+    </Compile>
     <Compile Include="..\common\XunitWorkerThread_Thread.cs">
       <Link>Common\XunitWorkerThread_Thread.cs</Link>
     </Compile>
