+++ src/xunit.execution.wp8/xunit.execution.wp8.csproj
@@ -128,6 +128,9 @@
     <Compile Include="..\common\WinRTFile.cs">
       <Link>Common\WinRTFile.cs</Link>
     </Compile>
+    <Compile Include="..\common\XunitSerializationInfo.cs">
+      <Link>Common\XunitSerializationInfo.cs</Link>
+    </Compile>
     <Compile Include="..\common\XunitWorkerThread_ThreadPool.cs">
       <Link>Common\XunitWorkerThread_ThreadPool.cs</Link>
     </Compile>
