+++ src/xunit.execution/Sdk/Reflection/ReflectionAttributeInfo.cs
@@ -51,7 +51,7 @@ namespace Xunit.Sdk
         {
             return attributeType.GetTypeInfo().GetCustomAttributes(typeof(AttributeUsageAttribute), true)
                                 .Cast<AttributeUsageAttribute>()
-                                .SingleOrDefault()
+                                .FirstOrDefault()
                 ?? DefaultAttributeUsageAttribute;
         }
 
@@ -96,8 +96,8 @@ namespace Xunit.Sdk
         /// <inheritdoc/>
         public TValue GetNamedArgument<TValue>(string propertyName)
         {
-            PropertyInfo propInfo = Attribute.GetType().GetRuntimeProperties().Single(pi => pi.Name == propertyName);
-            Guard.ArgumentValid("propertyName", "Could not find property " + propertyName + " on instance of " + Attribute.GetType().FullName, propInfo != null);
+            var propInfo = Attribute.GetType().GetRuntimeProperties().FirstOrDefault(pi => pi.Name == propertyName);
+            Guard.ArgumentValid("propertyName", String.Format("Could not find property {0} on instance of {1}", propertyName, Attribute.GetType().FullName), propInfo != null);
 
             return (TValue)propInfo.GetValue(Attribute, new object[0]);
         }
