+++ src/xunit.runner.utility.ios-universal/xunit.runner.utility.ios-universal.csproj
@@ -80,6 +80,9 @@
     <Compile Include="..\common\RemoteAppDomainManager_NoAppDomain.cs">
       <Link>Common\RemoteAppDomainManager_NoAppDomain.cs</Link>
     </Compile>
+    <Compile Include="..\common\SerializationHelper.cs">
+      <Link>Common\SerializationHelper.cs</Link>
+    </Compile>
     <Compile Include="..\common\SourceInformation.cs">
       <Link>Common\SourceInformation.cs</Link>
     </Compile>
@@ -97,6 +100,9 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\common\XunitSerializationInfo.cs">
+      <Link>Common\XunitSerializationInfo.cs</Link>
+    </Compile>
     <Compile Include="..\xunit.runner.utility\**\*.cs">
       <Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
     </Compile>
