+++ src/xunit.runner.utility.universal/xunit.runner.utility.universal.csproj
@@ -126,6 +126,9 @@
     <Compile Include="..\common\RemoteAppDomainManager_WPA81.cs">
       <Link>Common\RemoteAppDomainManager_WPA81.cs</Link>
     </Compile>
+    <Compile Include="..\common\SerializationHelper.cs">
+      <Link>Common\SerializationHelper.cs</Link>
+    </Compile>
     <Compile Include="..\common\SourceInformation.cs">
       <Link>Common\SourceInformation.cs</Link>
     </Compile>
@@ -147,8 +150,8 @@
     <Compile Include="..\common\WinRTFile.cs">
       <Link>Common\WinRTFile.cs</Link>
     </Compile>
-    <Compile Include="..\xunit.execution\Serialization\XunitSerializationInfo.cs">
-      <Link>Serialization\XunitSerializationInfo.cs</Link>
+    <Compile Include="..\common\XunitSerializationInfo.cs">
+      <Link>Common\XunitSerializationInfo.cs</Link>
     </Compile>
     <Compile Include="..\xunit.runner.utility\**\*.cs">
       <Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
