+++ src/xunit.runner.utility.wp8/xunit.runner.utility.wp8.csproj
@@ -129,6 +129,9 @@
     <Compile Include="..\common\RemoteAppDomainManager_WPA81.cs">
       <Link>Common\RemoteAppDomainManager_WPA81.cs</Link>
     </Compile>
+    <Compile Include="..\common\SerializationHelper.cs">
+      <Link>Common\SerializationHelper.cs</Link>
+    </Compile>
     <Compile Include="..\common\SourceInformation.cs">
       <Link>Common\SourceInformation.cs</Link>
     </Compile>
@@ -144,8 +147,8 @@
     <Compile Include="..\common\TestOptionsNames.cs">
       <Link>Common\TestOptionsNames.cs</Link>
     </Compile>
-    <Compile Include="..\xunit.execution\Serialization\XunitSerializationInfo.cs">
-      <Link>Serialization\XunitSerializationInfo.cs</Link>
+    <Compile Include="..\common\XunitSerializationInfo.cs">
+      <Link>Common\XunitSerializationInfo.cs</Link>
     </Compile>
     <Compile Include="..\xunit.runner.utility\**\*.cs">
       <Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
